Program compilation rework
