+++ Compilers/CPU_Compiler/Compile.bat
@@ -1,0 +1,1 @@
+g++ Main.cpp -o CPU_Compiler.exe -static-libgcc -static-libstdc++
