+++ Compilers/CPU_Compiler/Test.cbc
@@ -1,0 +1,13 @@
+5
+100
+34
+0
+0
+0
+1
+16
+0
+0
+0
+19
+0
